pyproject.toml
[project]
name = "adaptive-reasoning-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "fastapi==0.115.0",
    "gradio==4.44.1",
    "langchain>=1.0.0,<2.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=1.0.0a1,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=0.1.0",
    "langgraph>=0.2.28",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "sentence-transformers==3.1.1",
    "tavily-python==0.5.0",
    "uvicorn==0.31.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.3.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | chroma_db
)/
'''
